+++ internal/command/testdata/test/null_output/main.tf
@@ -1,0 +1,3 @@
+output "my_null_output" {
+  value = null
+}